combine utils and rename
