+++ oceanics-io-www/src/components/References/Reference.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import type {Meta, Story} from '@storybook/react';
-import {createIndex, readIndexedDocuments} from "../../../src/next-util";
+import {createIndex, readIndexedDocuments} from "../../shared";
 
 /**
  * Base component
