+++ oceanics-io-www/src/hooks/useObjectStorage.tsx
@@ -7,7 +7,7 @@ import { useEffect, useState } from "react";
  * Dedicated Worker loader
  */
 import type {FileSystem} from "../workers/shared";
-import type {WorkerRef} from "../utils";
+import type {WorkerRef} from "../shared";
 /**
  * The `useObjectStorage` hook provides a directory like structure
  * that describes assets in an S3-compatible storage service.
