--- oceanics-io-www/src/typography.ts
@@ -1,5 +1,0 @@
-import Typography from "typography";
-const typography = new Typography({});
-
-export default typography;
-export const {rhythm, scale} = typography;
